requirements.txt
pynetbox==6.1.3
python-gitlab==2.10.1
cryptography==3.0
napalm==3.2.0
